tests/wpt/tests/intersection-observer/scroll-margin-horizontal.html

<!DOCTYPE html>
<meta name="viewport" content="width=device-width,initial-scale=1">
<link rel="help" href="https://www.w3.org/TR/intersection-observer/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
#scroller { width: 100px; height: 100px; overflow: hidden; background-color: gray; display: flex; }
#spacer { width: 100px; height: 50px; flex-shrink: 0; }
#target { width: 50px; height: 50px; background-color: green; flex-shrink: 0; }
</style>

<div id=scroller>
  <div id=spacer></div>
  <div id=target></div>
</div>

<script>
let entries = [];

window.onload = function() {
  runTestCycle(testIntersection, "Test scroll margin intersection");

  const observer = new IntersectionObserver(
    es => entries = entries.concat(es),
    {
      scrollMargin: "10px"
    }
  );
  observer.observe(target);
};

function testIntersection() {
  assert_equals(entries.length, 1, "IntersectionObserverEntryCount");
  assert_true(entries[0].isIntersecting, "isIntersecting");
  assert_approx_equals(entries[0].intersectionRatio, 0.2, 0.001, "intersectionRatio");
}
</script>
